v0.5.6: Use float numbers for cost (instead of previous int), remove cost unit concept

+++ Source/Internal/Graph.cpp
@@ -19,7 +19,7 @@ namespace QDPF
 
 		void SimpleDirectedGraph::Init() {}
 
-		void SimpleDirectedGraph::AddEdge(int u, int v, int cost)
+		void SimpleDirectedGraph::AddEdge(int u, int v, float cost)
 		{
 			edges[u].insert({ v, cost });
 			predecessors[v].insert(u);
